requirements.txt
torch
transformers
pillow
torchvision
accelerate
torch-pca
torch-linear-assignment